Conveniently restart gpg-agent
I happen to have some random trouble with the yubikey: gpg-agent or
scdaemon seems to lose, but not forget about, the connection with the
card.  When this happens, all I see is "conflicting use".  Reloading the
agent works, but restarting emacs when this happens is a bore.  The
given script and function make the workaround somewhat less inconvenient.